requirements.txt
# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Azure OpenAI
openai>=1.50.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
